-Read 'default.aqu' from SD card (if present) and set timers accordingly during setup().
-dimmerTask now calculates and sets the LED duty cycle.

-New bool variable 'programOverride' to override the dimmerTask. (duh)

-mapFloat() now uses floats instead of doubles.

-Added api call 'aquacontrol32/api/setpercentage?percentage=x.xx' to set percentage to x.xx percent.

-Reduced stack sizes for all tasks.

-Some scheduling changes.

+++ routines.ino
@@ -1,4 +1,4 @@
-float mapFloat( double x, const double in_min, const double in_max, const double out_min, const double out_max)
+float mapFloat( float x, const float in_min, const float in_max, const float out_min, const float out_max)
 {
   return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
